修改 minist 数据集的训练代码

+++ samples.h
@@ -5,15 +5,15 @@ typedef struct {
     int num_samples;
     int num_input;
     int num_output;
-    double *data;
+    float *data;
 } SAMPLES;
 
 SAMPLES* samples_create    (int sampn, int inputn, int outputn);
 void     samples_destroy   (SAMPLES *samples);
 SAMPLES* samples_load      (char *file);
 void     samples_save      (SAMPLES *samples, char *file);
-double*  samples_get_input (SAMPLES *samples, int idx);
-double*  samples_get_output(SAMPLES *samples, int idx);
+float*   samples_get_input (SAMPLES *samples, int idx);
+float*   samples_get_output(SAMPLES *samples, int idx);
 
 #endif
